I wanted to move the leadershipboard file to utils since it is just a small part and is really just a util to start the leadership job
